add best practice for news api
