+++ News/urls.py
@@ -2,9 +2,5 @@ from  django.urls import path, include
 from . import views
 
 urlpatterns = [
-    path('get/all_news/',views.allNews.as_view(),name='allNews'),
-    path('add_news/',views.addNews.as_view(),name='addNews'),
-    path('delete_news/',views.deleteNews.as_view(),name='deleteNews'),
-    path('edit_news/',views.editNews.as_view(),name='editNews'),
-    path('api/news/',views.NewsApi.as_view(),name='newsApi')
+    path('api/',views.NewsApi.as_view(),name='newsApi')
 ]
